+++ packages/gamut/src/Modals/Modal.tsx
@@ -1,5 +1,5 @@
 import { MiniDeleteIcon } from '@codecademy/gamut-icons';
-import { ComponentProps, useRef, useState } from 'react';
+import { ComponentProps, useState } from 'react';
 import * as React from 'react';
 
 import { Box } from '../Box';
@@ -80,7 +80,6 @@ export const Modal: React.FC<ModalProps> = ({
   closeDisabled,
   ...rest
 }) => {
-  const buttonRef = useRef<HTMLButtonElement | null>(null);
   const [currentView, setCurrentView] = useState(0);
   const image = (views?.[currentView].image || rest?.image) ?? null;
 
@@ -92,14 +91,16 @@ export const Modal: React.FC<ModalProps> = ({
       {...rest}
     >
       <ModalContainer
-        className={className}
-        aria-label={ariaLabel}
-        aria-hidden="false"
-        aria-modal="true"
+        tabIndex={-1}
+        size={size}
         role="dialog"
         layout={views && views?.length > 0 ? 'dialog' : 'standard'}
-        size={size}
+        data-autofocus
+        className={className}
+        aria-modal="true"
         aria-live={ariaLive}
+        aria-label={ariaLabel}
+        aria-hidden="false"
       >
         {(title || views?.[currentView].title) && (
           <Text
@@ -119,7 +120,6 @@ export const Modal: React.FC<ModalProps> = ({
               icon={MiniDeleteIcon}
               onClick={onRequestClose}
               disabled={closeDisabled}
-              ref={buttonRef}
               tip="Close modal"
             />
           </Box>
@@ -150,7 +150,6 @@ export const Modal: React.FC<ModalProps> = ({
             variant="primary"
             onClick={(e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
               setCurrentView(currentView + 1);
-              buttonRef.current?.focus();
               views?.[currentView].nextCta?.onClick?.(e);
             }}
             gridArea="confirm"