+++ packages/gamut/src/Tip/InfoTip/elements.tsx
@@ -1,0 +1,8 @@
+import { css } from '@codecademy/gamut-styles';
+import styled from '@emotion/styled';
+
+import { Text } from '../../Typography';
+
+export const ScreenreaderNavigableTaxt = styled(Text)(
+  css({ position: 'relative' })
+);