+++ packages/gamut/src/Tip/ToolTip/index.tsx
@@ -69,7 +69,8 @@ export const ToolTip: React.FC<ToolTipProps> = ({
   return (
     <>
       {shouldRenderAriaTip && (
-        <Text screenreader id={id} role="tooltip">
+        // These are aria-hidden to ensure there's no duplication of content for screen readers navigating with CTRL + OPTION + ARROW
+        <Text aria-hidden screenreader id={id} role="tooltip">
           {adjustedInfo}
         </Text>
       )}