+++ packages/gamut-kit/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.6.427](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.426...@codecademy/gamut-kit@0.6.427) (2024-07-23)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+### [0.6.426](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.425...@codecademy/gamut-kit@0.6.426) (2024-07-22)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+### [0.6.425](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.424...@codecademy/gamut-kit@0.6.425) (2024-07-16)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
 ### [0.6.424](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.423...@codecademy/gamut-kit@0.6.424) (2024-07-08)
 
 **Note:** Version bump only for package @codecademy/gamut-kit